+++ src/Models/Auth.php
@@ -6,7 +6,6 @@ class Auth
 {
     function __construct(
         public $userId,
-        public $apiKey
     )
     {